pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-code-review-env"
version = "0.3.0"
description = "Deterministic Python code review and repair benchmark environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.40.0",
    "pytest>=8.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"
python-code-review-env-server = "python_code_review_env.envs.python_env_env.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["python_code_review_env*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
